Revert "[17기_서찬혁] DRF2 - Simple JWT 미션 제출합니다"
