--- .env.prod
@@ -1,9 +1,0 @@
-DATABASE_HOST=database-1.chikj49ulvtb.ap-northeast-2.rds.amazonaws.com
-DATABASE_DB=mysql
-DATABASE_NAME=database-1
-DATABASE_USER=root
-DATABASE_PASSWORD=13241324
-DATABASE_PORT=3306
-DEBUG=False
-DJANGO_ALLOWED_HOSTS=ec2-3-34-91-156.ap-northeast-2.compute.amazonaws.com
-DJANGO_SECRET_KEY=roltvk+e-r5lk)6tnn+(npum)c5a0a*j5weag891rc^i^nrx!+
