+++ .github/workflows/deploy.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - dev
-      - master
 
 jobs:
