+++ account/admin.py
@@ -1,41 +1,8 @@
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
-from django.contrib.auth.models import Group
-from django.utils.translation import gettext_lazy as _
 
-from .forms import UserCreationForm, UserChangeForm
 from .models import User, UserSubject
+# Register your models here.
 
-
-class UserAdmin(BaseUserAdmin):
-    # The forms to add and change user instances
-    form = UserChangeForm
-    add_form = UserCreationForm
-
-    # The fields to be used in displaying the User model.
-    # These override the definitions on the base UserAdmin
-    # that reference specific fields on auth.User.
-    list_display = ('email', 'username', 'is_active', 'is_superuser', 'date_joined')
-    list_filter = ('is_superuser', 'is_active',)
-    fieldsets = (
-        (None, {'fields': ('email', 'password')}),
-        (_('Personal info'), {'fields': ('username', )}),
-        (_('Permissions'), {'fields': ('is_active', 'is_superuser',)}),
-    )
-    # add_fieldsets is not a standard ModelAdmin attribute. UserAdmin
-    # overrides get_fieldsets to use this attribute when creating a user.
-    add_fieldsets = (
-        (None, {
-            'classes': ('wide',),
-            'fields': ('email', 'username', 'password1', 'password2')}
-         ),
-    )
-    search_fields = ('email', 'username')
-    ordering = ('-date_joined',)
-    filter_horizontal = ()
-
-
-# Now register the new UserAdmin...
-admin.site.register(User, UserAdmin)
+admin.site.register(User)
 admin.site.register(UserSubject)
