+++ account/models/__init__.py
@@ -2,5 +2,5 @@ from django.db import models
 
 # Create your models here.
 
-from .user import User, UserManager
+from .user import User
 from .user_subject import UserSubject
