+++ account/models/user.py
@@ -1,95 +1,13 @@
-from django.contrib.auth.models import (
-    BaseUserManager, AbstractBaseUser, PermissionsMixin
-)
 from django.db import models
-from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
+
+from django.contrib.auth.models import AbstractUser
 
 
-
-class UserManager(BaseUserManager):
-    def create_user(self, email, username, password=None):
-        """
-        주어진 이메일, 닉네임, 비밀번호 등 개인정보로 User 인스턴스 생성
-        """
-        if not email:
-            raise ValueError(_('Users must have an email address'))
-
-        user = self.model(
-            email=self.normalize_email(email),
-            nickname=username,
-        )
-
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, email, username, last_name, first_name, password):
-        """
-        주어진 이메일, 닉네임, 비밀번호 등 개인정보로 User 인스턴스 생성
-        단, 최상위 사용자이므로 권한을 부여한다.
-        """
-        user = self.create_user(
-            email=email,
-            password=password,
-            username=username,
-        )
-
-        user.is_superuser = True
-        user.save(using=self._db)
-        return user
-
-
-class User(AbstractBaseUser, PermissionsMixin):
-    email = models.EmailField(
-        verbose_name=_('Email address'),
-        max_length=255,
-        unique=True,
-    )
-    username = models.CharField(
-        verbose_name=_('Username'),
-        max_length=30,
-        unique=True
-    )
-    is_active = models.BooleanField(
-        verbose_name=_('Is active'),
-        default=True
-    )
-    date_joined = models.DateTimeField(
-        verbose_name=_('Date joined'),
-        default=timezone.now
-    )
-    first_name = models.CharField(
-        verbose_name=_('First name'),
-        max_length=30,
-        blank=True
-    )
-    last_name = models.CharField(
-        verbose_name=_('Last name'),
-        max_length=30,
-        blank=True
-    )
-    profile_image = models.ImageField(
-        verbose_name=_('Profile image'),
-        upload_to='profile_image',
-        blank=True
-    )
-
-    objects = UserManager()
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = ['username', ]
-
-    class Meta:
-        verbose_name = _('user')
-        verbose_name_plural = _('users')
-        ordering = ('-date_joined',)
+class User(AbstractUser):
+    friend_list = models.ManyToManyField('self', blank=True)
+    profile_image = models.ImageField(upload_to='profile_image', blank=True, null=True)
 
     def __str__(self):
         return self.username
 
-    @property
-    def is_staff(self):
-        "Is the user a member of staff?"
-        # Simplest possible answer: All superusers are staff
-        return self.is_superuser
+
