+++ account/serializers/user_serializer.py
@@ -7,4 +7,4 @@ User = get_user_model()
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
-        fields = '__all__'
+        fields = ('id', 'username', 'email', 'first_name', 'last_name', 'profile_image', 'friend_list')
