--- account/views/__init__.py
@@ -1,4 +1,0 @@
-from .register_view import RegisterAPIView
-from .user_list_view import UserListAPIView
-from .login_view import LoginAPIView
-from .logout_view import LogoutAPIView
