--- account/views/login_view.py
@@ -1,36 +1,0 @@
-from rest_framework import status
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from account.serializers import UserSerializer
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from django.contrib.auth import authenticate
-
-
-class LoginAPIView(APIView):
-
-    @staticmethod
-    def post(request):
-        user = authenticate(
-            email=request.data.get("email"), password=request.data.get("password")
-        )
-        if user is not None:
-            serializer = UserSerializer(user)
-            token = TokenObtainPairSerializer.get_token(user)
-            refresh_token = str(token)
-            access_token = str(token.access_token)
-            res = Response(
-                {
-                    "user": serializer.data,
-                    "message": "login success",
-                    "token": {
-                        "access": access_token,
-                        "refresh": refresh_token,
-                    },
-                },
-                status=status.HTTP_200_OK,
-            )
-            res.set_cookie("access", access_token, httponly=True)
-            res.set_cookie("refresh", refresh_token, httponly=True)
-            return res
-        else:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
