--- account/views/logout_view.py
@@ -1,19 +1,0 @@
-from rest_framework import status
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework.permissions import IsAuthenticated
-
-
-class LogoutAPIView(APIView):
-    permission_classes = (IsAuthenticated,)
-
-    @staticmethod
-    def post(request):
-        try:
-            refresh_token = request.COOKIES.get("refresh")
-            token = RefreshToken(refresh_token)
-            token.blacklist()
-            return Response(status=status.HTTP_205_RESET_CONTENT)
-        except Exception as e:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
