--- account/views/register_view.py
@@ -1,33 +1,0 @@
-from rest_framework import status
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from account.serializers import RegisterSerializer
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-
-
-class RegisterAPIView(APIView):
-
-    @staticmethod
-    def post(request):
-        serializer = RegisterSerializer(data=request.data)
-        if serializer.is_valid():
-            user = serializer.save()
-            # jwt token 접근해주기
-            token = TokenObtainPairSerializer.get_token(user)
-            refresh_token = str(token)
-            access_token = str(token.access_token)
-            res = Response(
-                {
-                    "user": serializer.data,
-                    "message": "register successs",
-                    "token": {
-                        "access": access_token,
-                        "refresh": refresh_token,
-                    },
-                },
-                status=status.HTTP_200_OK,
-            )
-            res.set_cookie("access", access_token, httponly=True)
-            res.set_cookie("refresh", refresh_token, httponly=True)
-            return res
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
