+++ django_rest_framework_17th/settings/dev.py
@@ -1,4 +1,6 @@
+import pymysql
 from .base import *  # noqa
+pymysql.install_as_MySQLdb()
 
 DEBUG = True
