+++ django_rest_framework_17th/settings/prod.py
@@ -1,6 +1,5 @@
 from .base import *  # noqa
 
-
 DEBUG = False
 ALLOWED_HOSTS = env.list('DJANGO_ALLOWED_HOSTS')
