+++ docker-compose.prod.yml
@@ -3,31 +3,20 @@ services:
 
   web:
     container_name: web
-    build:
-      context: ./
-      dockerfile: Dockerfile.prod
-    command: gunicorn django_rest_framework_17th.wsgi:application --bind 0.0.0.0:8000
-    environment:
-      DJANGO_SETTINGS_MODULE: django_rest_framework_17th.settings.prod
-    env_file:
-      - ./.env.prod
+    #작성
     volumes:
       - static:/home/app/web/static
       - media:/home/app/web/media
-    expose:
-      - 8000
     entrypoint:
       - sh
       - config/docker/entrypoint.prod.sh
 
   nginx:
     container_name: nginx
-    build: ./config/nginx
+    #작성
     volumes:
       - static:/home/app/web/static
       - media:/home/app/web/media
-    ports:
-      - 80:80
     depends_on:
       - web
