+++ docker-compose.yml
@@ -3,35 +3,13 @@ services:
 
   db:
     container_name: db
-    image: mariadb:latest #mac
-    restart: always
-    environment:
-      MYSQL_ROOT_HOST: '%'
-      MYSQL_ROOT_PASSWORD: mysql
-    expose:
-      - 3306
-    ports:
-      - "3307:3306"
-    env_file:
-      - .env
+    #작성
     volumes:
       - dbdata:/var/lib/mysql
 
   web:
     container_name: web
-    build: .
-    command: sh -c "python manage.py migrate && python manage.py runserver 0.0.0.0:8000"
-    environment:
-      MYSQL_ROOT_PASSWORD: mysql
-      DATABASE_NAME: mysql
-      DATABASE_USER: 'root'
-      DATABASE_PASSWORD: mysql
-      DATABASE_PORT: 3306
-      DATABASE_HOST: db
-      DJANGO_SETTINGS_MODULE: django_rest_framework_17th.settings.dev
-    restart: always
-    ports:
-      - "8000:8000"
+    #작성
     volumes:
       - .:/app
     depends_on:
