+++ requirements.txt
@@ -1,11 +1,9 @@
-asgiref==3.6.0
-Django==4.2.1
-django-environ==0.10.0
-django-filter==23.2
-djangorestframework==3.14.0
-djangorestframework-simplejwt==5.2.2
-gunicorn==20.1.0
-mysqlclient==2.1.1
-Pillow==9.5.0
-pytz==2023.3
-sqlparse==0.4.4
+asgiref==3.3.2
+Django==3.2.16
+django-environ==0.4.5
+gunicorn==20.0.4
+pytz==2020.1
+sqlparse==0.3.1
+PyMySQL~=1.0.3
+djangorestframework~=3.14.0
+django-filter~=23.1
